fix(activate): rename activate to promote
